pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "baton"
version = "0.1.0"
description = "Multi-project dashboard for scaffold-based projects"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "httpx>=0.25",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[project.scripts]
baton-agent = "backend.agent:main"

[project.optional-dependencies]
dev = ["ruff", "pytest", "httpx"]
